Version1/requirements.txt
langgraph>=0.0.55
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
openai>=1.10.0
sqlalchemy>=2.0.23
pandas>=2.1.3
plotly>=5.17.0
streamlit>=1.28.1
psycopg>=3.1.0
python-dotenv>=1.0.0
qdrant-client>=1.6.0
pymupdf>=1.23.0
tiktoken>=0.5.0